moved draw() from MarkerManager to AbstractMarkerManager, accidentally added markers to AbstractMarkerManagers before, fixed now
